requirements.txt
# ML
numpy==1.20.3

# DATA ANALYSIS
matplotlib==3.4.2
nnfs~=0.5.1
scikit-learn~=0.24.2
pandas~=1.2.4
tqdm~=4.61.1
Pillow~=8.2.0
idx2numpy~=1.2.3